+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Change log itself follows [Keep a CHANGELOG](http://keepachangelog.com) format.
 ### Added
 
 * `Faker.Superhero` [@orieken][]
+* `Faker.Code.iban` [@tobyhinloopen][]
 * `Faker.Beer` [@orieken][]
 * `Faker.Date` [@tobyhinloopen][]
 * `Faker.Nato` [@petehamilton][]